Fix transaction issue
