+++ apps/question-bank/src/main/java/gov/cdc/nbs/questionbank/phin/VocabSearchController.java
@@ -1,11 +1,7 @@
-package gov.cdc.nbs.questionbank.question;
+package gov.cdc.nbs.questionbank.phin;
 
-import gov.cdc.nbs.questionbank.question.response.Status;
-import gov.cdc.nbs.questionbank.question.response.ValueSetByOIDResponse;
-import gov.cdc.nbs.questionbank.question.response.ValueSetByOIDResults;
-import gov.cdc.nbs.questionbank.question.service.VocabSearchService;
-import lombok.NonNull;
-import lombok.extern.slf4j.Slf4j;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,11 +10,12 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
+import gov.cdc.nbs.questionbank.phin.service.VocabSearchService;
+import gov.cdc.nbs.questionbank.question.response.Status;
+import gov.cdc.nbs.questionbank.question.response.ValueSetByOIDResponse;
+import gov.cdc.nbs.questionbank.question.response.ValueSetByOIDResults;
+import lombok.NonNull;
 
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
-@Slf4j
 @RestController
 @RequestMapping("/api/v1/phin")
 public class VocabSearchController {
@@ -30,7 +27,7 @@ public class VocabSearchController {
     ResponseEntity<ValueSetByOIDResponse> fetchValueSetInfoByOID(@PathVariable("oid") @NonNull String oid) {
         ValueSetByOIDResponse response = null;
         Status status = null;
-        if(isValidInput(oid)) {
+        if (isValidInput(oid)) {
             try {
                 ValueSetByOIDResults data = vocabSearchService.fetchValueSetInfoByOID(oid);
                 status = Status.builder().code("200").type("SUCCESS").message("OID_DATA_FOUND")
@@ -53,9 +50,9 @@ public class VocabSearchController {
                 .description("Invalid input").build();
         response = ValueSetByOIDResponse.builder().status(status).build();
         return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
-  }
+    }
 
-    private boolean isValidInput(String oid){
+    private boolean isValidInput(String oid) {
         oid = oid.replace(".", "");
         Pattern pattern = Pattern.compile("[0-9]+");
         Matcher matcher = pattern.matcher(oid);
@@ -64,6 +61,5 @@ public class VocabSearchController {
 
 
 }
-
