+++ apps/question-bank/src/main/java/gov/cdc/nbs/questionbank/phin/client/PhinVadsClient.java
@@ -1,14 +1,13 @@
-package gov.cdc.nbs.questionbank.question.client;
+package gov.cdc.nbs.questionbank.phin.client;
 
-import gov.cdc.nbs.questionbank.question.response.ValueSetApiResponse;
-import gov.cdc.nbs.questionbank.question.response.PhinvadsValueSetByIDData;
-import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.web.server.ResponseStatusException;
+import gov.cdc.nbs.questionbank.question.response.PhinvadsValueSetByIDData;
+import lombok.extern.slf4j.Slf4j;
 
 @Component
 @Slf4j
@@ -17,7 +16,7 @@ public class PhinVadsClient {
     @Value("${valueset.byidurl}")
     private String valuesetByIDUrl;
 
-    public PhinvadsValueSetByIDData getValueSetByOID(String oid) throws  ResponseStatusException{
+    public PhinvadsValueSetByIDData getValueSetByOID(String oid) throws ResponseStatusException {
         RestTemplate restTemplate = new RestTemplate();
         ResponseEntity<PhinvadsValueSetByIDData> clientResp = restTemplate.getForEntity(valuesetByIDUrl + oid,
                 PhinvadsValueSetByIDData.class);
