+++ apps/question-bank/src/test/java/gov/cdc/nbs/questionbank/phin/service/VocabSearchServiceTest.java
@@ -1,29 +1,23 @@
-package gov.cdc.nbs.questionbank.question.service;
+package gov.cdc.nbs.questionbank.phin.service;
 
-import gov.cdc.nbs.questionbank.question.VocabSearchController;
-import gov.cdc.nbs.questionbank.question.client.PhinVadsClient;
-import gov.cdc.nbs.questionbank.question.model.Compose;
-import gov.cdc.nbs.questionbank.question.model.Concept;
-import gov.cdc.nbs.questionbank.question.model.Include;
-import gov.cdc.nbs.questionbank.question.response.PhinvadsValueSetByIDData;
-import gov.cdc.nbs.questionbank.question.response.ValueSetByOIDResponse;
-import gov.cdc.nbs.questionbank.question.response.ValueSetByOIDResults;
-import gov.cdc.nbs.questionbank.question.response.ValueSetConcept;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.Mockito.when;
+import java.util.ArrayList;
+import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.server.ResponseStatusException;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.Mockito.when;
+import gov.cdc.nbs.questionbank.phin.client.PhinVadsClient;
+import gov.cdc.nbs.questionbank.question.model.Compose;
+import gov.cdc.nbs.questionbank.question.model.Concept;
+import gov.cdc.nbs.questionbank.question.model.Include;
+import gov.cdc.nbs.questionbank.question.response.PhinvadsValueSetByIDData;
+import gov.cdc.nbs.questionbank.question.response.ValueSetByOIDResults;
 
 @ExtendWith(MockitoExtension.class)
 class VocabSearchServiceTest {
@@ -60,7 +54,7 @@ class VocabSearchServiceTest {
     void should_return_exception_when_oid_not_found() {
         when(phinVadsClient.getValueSetByOID("2.16.840.1.114222.4.11.000")).thenThrow(ResponseStatusException.class);
         try {
-            ValueSetByOIDResults results = vocabSearchService.fetchValueSetInfoByOID("2.16.840.1.114222.4.11.000");
+            vocabSearchService.fetchValueSetInfoByOID("2.16.840.1.114222.4.11.000");
         }catch (ResponseStatusException e){
             assertTrue(e instanceof  ResponseStatusException);
         }
