+++ apps/question-bank/src/test/java/gov/cdc/nbs/questionbank/question/UpdateQuestionSteps.java
@@ -11,7 +11,7 @@ import gov.cdc.nbs.questionbank.question.exception.QuestionNotFoundException;
 import gov.cdc.nbs.questionbank.question.repository.WaQuestionRepository;
 import gov.cdc.nbs.questionbank.question.request.QuestionType;
 import gov.cdc.nbs.questionbank.question.request.UpdateQuestionRequest;
-import gov.cdc.nbs.questionbank.question.util.ExceptionHolder;
+import gov.cdc.nbs.questionbank.support.ExceptionHolder;
 import gov.cdc.nbs.questionbank.support.QuestionMother;
 import gov.cdc.nbs.questionbank.support.QuestionRequestMother;
 import io.cucumber.java.en.Then;
